+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+  - 💄 [@mixte/mel-components] 组件 `MelSelect` 的 `options` 选项支持 `MaybeRefOrGetter` 类型
 
 ## [v3.5.0-beta.15]
   - 📅 2025-08-31